prikbord gedeelte gemaakt, werkt 97%

+++ src/main/java/nl/natuurverhaal/natuurverhaal/controllers/ExcursieController.java
@@ -64,7 +64,7 @@ public class ExcursieController {
                                                                 @RequestPart("niveau") String niveau,
                                                                 @RequestPart("guide") String guide,
                                                                 @RequestPart("content") String content,
-                                                                @RequestPart("max_participants") Short max_participants)
+                                                                @RequestPart("max_participants") int max_participants)
 
                 throws IOException {
             ObjectMapper objectMapper = new ObjectMapper();
@@ -90,7 +90,7 @@ public class ExcursieController {
             excursie.setNiveau(niveau);
             excursie.setGuide(guide);
             excursie.setContent(content);
-            excursie.setMax_participants(max_participants);
+            excursie.setMax_participants((short)max_participants);
             excursie.setDate(formattedDateTime);
             OutputExcursieDto createdExcursie = excursieService.createExcursie(excursie);
             return ResponseEntity.status(HttpStatus.CREATED).body(createdExcursie);
